Don't try to thumbnail images that don't exist
Ideally this would be addressed by adopting the `default` option
available in the `easy_thumbnails` `thumbnail` tag. For now, this at
least prevents errors being thrown when thumbnailing images that don't
exist.

Also re-named one of the test classes to avoid a naming conflict (the
template tag tests weren't being run because the test runner was
ignoring the other test class with the same name).