Make django-image-cropping compatible with non-default file-storage
Previously, using ManifestStaticFilesStorage with django-
image-cropping wouldn't work, because the path to the file
was hard-coded. This change cribs from django.contrib.admin
for manipulating the path to the static file before handing
it off to Django forms. Arguably, it'd be nice if this is
something Django handled for us, but until it does, we need
to do this work ourselves.

For Django versions earlier than 1.7, we don't change
behaviour, since figuring out whether the staticfiles app is
installed is hard without django.apps.